tweaked my level further
lots of changes and added an experimental shortcut block. could make that one a bit more interesting but now jumping over the pipe is significantly more rewarding than before. Maybe too much of a time save? It's a cool trick but it felt even cooler without the platform afterward, when you could bounce on a slope below to gain lots of speed.