Added more Alt layers
naming convention is
LayerAltN_000
so the first one is called LayerAlt1_001 and the next is LayerAlt2_001
the first number is based on which bg layer it is. The Zeros and a number at the end are for telling in what order they were made. So the second "Layer 2" I made is called LayerAlt2_002 for example. I kept numbering each separate Terrain with this in mind continously even in different scenes, so they never share the same name as another Terrain.